requirements.txt
vosk>=0.3.45
sounddevice>=0.4.6
numpy>=1.24.0
moviepy>=1.0.3
num2words>=0.5.12
arabic-reshaper>=3.0.0
python-bidi>=0.4.2
openai-whisper>=20231117
pyaudio>=0.2.13
